Refactor transducer proxy structure: separate internal and public interfaces, add SyncSuspendingProxy implementation
